python_server/requirements.txt
# for pip install -r requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3